commands: move diffedit code to diffedit.rs
